Add checkpoint support for `GreatExpectationsBigQueryOperator`
This commit adds the checkpoint support for `GreatExpectationsBigQueryOperator`.
Currently this has not been working since `expectation_suite_name` was listed
as a pos_arg (instead of a var_pos_arg), which made it mandatory, because the
operators use the `@apply_defaults` decorator.

I removed the pos_arg entirely, since the `self.expectation_suite_name` member
is set by the `super()` call. Another alternative would be to make it a kwarg.
In general, all airflow operators should prefer kwargs over pos_arg.

Additionally, I changed the check for mandatory kwargs in
`GreatExpectationsOperator` to the intended behaviour:

- `expectation_suite_name` and `batch_kwargs`, or
- `assets_to_validate`, or
- `checkpoint_name`

To achieve this, I added a "logical XOR" to the check -
`(bool(expectation_suite_name) != bool(batch_kwargs)) * 2` - which is only True
when only one of them is set. By multiplying it with 2 we exclude the chance of
**only** setting one of them and no other mandatory kwargs. Basically, if only
one of these two are set, it will fail (because the `sum()`) will be `!= 1`.