Update Operator on_failure_callback to match baseOperator
from airflow's [baseOperator documentation](https://airflow.apache.org/docs/apache-airflow/stable/_api/airflow/models/baseoperator/index.html): 
```
on_failure_callback (TaskStateChangeCallback) -- a function to be called when a task instance of this task fails. a context dictionary is passed as a single parameter to this function. Context contains references to related objects to the task instance and is documented under the macros section of the API.
```

So propose to change the GreatExpectationsOperator follow this