Improve Performance of GreatExpectationsOperator
Moved the data context and checkpoint initializations
from the __init__() method to the execute() method of
the operator. Because an operator's __init__() is called
every time a DAG is parsed, the data context was being
created on every parse, slowing down the DAGs this operator
was in significantly. This change ensures that the data
context is only created during operator execution, when
it is needed.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>