Rewrite datasource process
Datasources were not being created correctly from the runtime_environment
parameter passed into the DataContext constructor. The datasource did not
seem to get built (tried only with the SQL Datasource, but tried to build it
as a Legacy Datasource, BaseDatasource, SQLAlchemyDatasource, and
SimpleSqlalchemyDatasource, none of which worked for various reasons);
in the SimpleSqlalchemyDatasource, the issue seemed to be that the
execution_engine config passed to BaseDatasource did not contain the
necessary "name" parameter, causing instantiation in data_context.util
(ln 121) to fail.