Fix template extension bug (#64)
Adds a comma to the end of the template_ext set, making the quoted
'.sql' a member of the set, not each character in the string a member
of the set. This was breaking jinja templating. Testing has been
updated to ensure the whole value, not the characters, of the elements
of the template_ext set are present.