Add schema parameter and data_asset_name parsing
A new schema parameter will overwrite the supplied connection's
schema, if it exists. If a data_asset_name is provided in the form
'schema.table', the schema will be parsed out into self.schema and
data_asset_name will be updated to only be the table. This will also
override the supplied schema in the connection.