Add schema parameter and data_asset_name parsing (#75)
* Add schema parameter and data_asset_name parsing

A new schema parameter will overwrite the supplied connection's
schema, if it exists. If a data_asset_name is provided in the form
'schema.table', the schema will be parsed out into self.schema and
data_asset_name will be updated to only be the table. This will also
override the supplied schema in the connection.

* Fix new test name

* Update great_expectations_provider/operators/great_expectations.py

Co-authored-by: Kaxil Naik <kaxilnaik@gmail.com>

* Remove conn param as it was a test from a different branch.

Co-authored-by: Kaxil Naik <kaxilnaik@gmail.com>