Add data asset name guard (#85)
* Bugfix: Remove unnecessary ValueError

An extraneous ValueError was in build_runtime_datasources and caused
incorrect operator failure when a runtime datasource is built but
a connection or dataframe is not passed in. The if/elif/else block
would hit the else, which was a ValueError, and it made the below
if statement on checkpoint_kwargs unreachable. It has been removed
as the ValueError in the first if block is sufficient to halt
operator execution in the case of incorrect usage.

* Add check in init to ensure a data_asset_name is accompanied by a conn_id or dataframe. This prevents issues with building a custom data asset at runtime.