Add Athena Connection Support
Add an Athena URI builder to make_connection_string(),
assuming for now that Athena is the only connection when an AWS
connection type is given. This is an incorrect assumption, but we
currently do not have asks for other use cases figuring out how
to differentiate these may be a non-trivial issue.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>