Add Athena Connection Support (#91)
* Add Athena Connection Support

Add an Athena URI builder to make_connection_string(),
assuming for now that Athena is the only connection when an AWS
connection type is given. This is an incorrect assumption, but we
currently do not have asks for other use cases figuring out how
to differentiate these may be a non-trivial issue.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Use params instead of kwargs to pass Athena-specific parameters. Not sure if this is correct use of params...

Signed-off-by: Benji Lampel <benjamin@astronomer.io>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add Athena Connection Support

Add an Athena URI builder to make_connection_string(),
assuming for now that Athena is the only connection when an AWS
connection type is given. This is an incorrect assumption, but we
currently do not have asks for other use cases figuring out how
to differentiate these may be a non-trivial issue.

Signed-off-by: Benji Lampel <benjamin@astronomer.io>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Use params instead of kwargs to pass Athena-specific parameters. Not sure if this is correct use of params...

Signed-off-by: Benji Lampel <benjamin@astronomer.io>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fix flake8 errors

Signed-off-by: Benji Lampel <benjamin@astronomer.io>

* Fix tests

Signed-off-by: Benji Lampel <benjamin@astronomer.io>

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fix flake8 and black issues

Signed-off-by: Benji Lampel <benjamin@astronomer.io>

---------

Signed-off-by: Benji Lampel <benjamin@astronomer.io>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>