Set version statically to let verify_tag script pass (#172)
The publish package job failed due to recent changes in the init file
that tries to get the version from another module.


https://github.com/astronomer/airflow-provider-great-expectations/actions/runs/13457867015/job/37605752581

So, to help unblock the release efforts, I am statically copying back
the version to init file and we could rework on adjusting the verify tag
script to accommodate using version from another module.